backend/requirements.txt
selenium>=4.10.0
undetected-chromedriver==3.5.4
webdriver-manager>=3.8.6
python-dotenv>=1.0.0
flask>=2.0.0
flask-cors>=3.0.10
google-generativeai
# Remove incorrect logging package - it's part of Python's standard library
fpdf>=1.7.2
python-docx>=0.8.11
